Add fp-dynamic-fee crate
