+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
 	"client/db",
 	"client/mapping-sync",
 	"primitives/consensus",
+	"primitives/dynamic-fee",
 	"primitives/evm",
 	"primitives/rpc",
 	"primitives/self-contained",
