+++ primitives/dynamic-fee/CHANGELOG.md
@@ -1,0 +1,3 @@
+# Changelog for `fp-dynamic-fee`
+
+## Unreleased
